+++ e2e/tests-dfx/nns.bash
@@ -45,13 +45,78 @@ teardown() {
 }
 
 @test "dfx nns install command exists" {
-    dfx_start
+    assert_command dfx nns install --help
+}
 
-    # it just calls ic-nns-init, which panics
-    assert_command_fail dfx nns install
-    assert_match "thread 'main' panicked at 'Illegal arguments:"
-    assert_match "ic-nns-init \[OPTIONS\]"
-    assert_match "ic-nns-init call failed"
+# Tries to start dfx on the default port, repeating until it succeeds or times out.
+#
+# Motivation: dfx nns install works only on port 8080, as URLs are compiled into the wasms.  This means that multiple
+# tests MAY compete for the same port.
+# - It may be possible in future for the wasms to detect their own URL and recompute signatures accordingly,
+#   however until such a time, we have this restriction.
+# - It may also be that ic-nns-install, if used on a non-standard port, installs only the core canisters not the UI.
+# - However until we have implemented good solutions, all tests on ic-nns-install must run on port 8080.
+dfx_start_for_nns_install() {
+    # TODO: When nns-dapp supports dynamic ports, this wait can be removed.
+    assert_command timeout 300 sh -c \
+        "until dfx start --clean --background --host 127.0.0.1:8080; do echo waiting for port 8080 to become free; sleep 3; done" \
+        || (echo "could not connect to replica on port 8080" && exit 1)
+    assert_match "subnet_type: System"
+    assert_match "127.0.0.1:8080"
+}
+
+# The nns canisters should be installed without changing any of the developer's project files,
+# so we cannot rely on `dfx canister id` when testing.  We rely on these hard-wired values instead:
+nns_canister_id() {
+    case "$1" in
+    nns-registry)          echo "rwlgt-iiaaa-aaaaa-aaaaa-cai" ;;
+    nns-governance)        echo "rrkah-fqaaa-aaaaa-aaaaq-cai" ;;
+    nns-ledger)            echo "ryjl3-tyaaa-aaaaa-aaaba-cai" ;;
+    nns-root)              echo "r7inp-6aaaa-aaaaa-aaabq-cai" ;;
+    nns-cycles-minting)    echo "rkp4c-7iaaa-aaaaa-aaaca-cai" ;;
+    nns-lifeline)          echo "rno2w-sqaaa-aaaaa-aaacq-cai" ;;
+    nns-genesis-token)     echo "renrk-eyaaa-aaaaa-aaada-cai" ;;
+    nns-sns-wasm)          echo "qjdve-lqaaa-aaaaa-aaaeq-cai" ;;
+    internet_identity)     echo "qaa6y-5yaaa-aaaaa-aaafa-cai" ;;
+    nns-dapp)              echo "qhbym-qaaaa-aaaaa-aaafq-cai" ;;
+    *)                     echo "ERROR: Unknown NNS canister '$1'." >&2
+                           exit 1;;
+    esac
+}
+
+@test "dfx nns install runs" {
+    # Setup
+    install_shared_asset subnet_type/shared_network_settings/system
+    dfx_start_for_nns_install
+    dfx nns install
+
+    # Checking that the install worked.
+    # Note:  The installation is quite expensive, so we test extensively on one installation
+    #        rather than doing a separate installation for every test.  The tests are read-only
+    #        so no test should affect the output of another.
+    installed_wasm_hash() {
+        dfx canister info "$(nns_canister_id "$1")" | awk '/Module hash/{print $3; exit 0}END{exit 1}'
+    }
+    downloaded_wasm_hash() {
+        sha256sum ".dfx/wasms/nns/$(dfx --version | awk '{printf "%s-$%s", $1, $2}')/$1" | awk '{print "0x" $1}'
+    }
+    wasm_matches() {
+            [[ "$(installed_wasm_hash "$1")" == "$(downloaded_wasm_hash "$2")" ]] || {
+                echo "ERROR:  There is a wasm hash mismatch between $1 and $2"
+                exit 1
+            }>&2
+    }
+    wasm_matches nns-registry registry-canister.wasm
+    wasm_matches nns-governance governance-canister_test.wasm
+    wasm_matches nns-ledger ledger-canister_notify-method.wasm
+    wasm_matches nns-root root-canister.wasm
+    wasm_matches nns-cycles-minting cycles-minting-canister.wasm
+    wasm_matches nns-lifeline lifeline.wasm
+    wasm_matches nns-genesis-token genesis-token-canister.wasm
+    wasm_matches nns-sns-wasm sns-wasm-canister.wasm
+    wasm_matches internet_identity internet_identity_dev.wasm
+    wasm_matches nns-dapp nns-dapp_local.wasm
+    dfx stop
 }
 
 test_project_import() {